requirements.txt
tensorflow>=2.9.0
pandas
numpy
Pillow
opencv-python
